Rework calculations (#708)
* Rework calculations

* Rework calculations

* Experiment Details Report; ExperimentObject; refactor Experiment tests

* Rename mutate4 REST method

* Validate patch on backend

* Use EnteredValue.$overwritten flag

* Verify undo/redo on a backend

* Experiment.version

* Experiment Model Assertions

* Revert "Use EnteredValue.$overwritten flag"

This reverts commit ed418d6a

* Experiment Model Assertions

* Update field list in Indident report #697

* Reapply "Use EnteredValue.$overwritten flag"

This reverts commit 55b219f5d9339933eed57f9fd9f17a85fb743cf0.

# Conflicts:
#	backend/eln/eln-core/src/main/java/com/epam/indigoeln/reaction/service/calculator/ReactionCalculator.java

* $overwritten test

* Lessen WAF rule for incidents path

* Make EnteredValue immutable

* Keep exact value for compound mol weight

* Test cases

* Make EnteredValue non-nullable; EnteredValue.present and EnteredValue.overwritten

* Lock experiment for updates

* Experiment Signer role

* S3 storage

* Use FileStorage in IncidentReportService

* Fixes after merge

* Fix integration tests

* Incident report fields

* Incident report fields

+++ backend/eln/eln-core/src/main/java/com/epam/indigoeln/compound/entity/CompoundEntity.java
@@ -13,7 +13,6 @@ import org.hibernate.annotations.JdbcType;
 import org.hibernate.dialect.type.PostgreSQLEnumJdbcType;
 import org.jspecify.annotations.Nullable;
 
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -65,10 +64,10 @@ public class CompoundEntity extends IdentifiableEntity {
     private String formula;
 
     @NotNull
-    private BigDecimal molWeight;
+    private Double molWeight;
 
     @NotNull
-    private BigDecimal exactMass;
+    private Double exactMass;
 
     @NotEmpty
     @Basic(fetch = FetchType.LAZY)
